Achieved MVP having impl. TextParser
Previous commit worked towards implementing the logic around filtering and storing the post filtered sentences. In conjunction with this commit, this application is currently a MVP (minimal viable product). It correctly filters and displays contents of said filtered text to the console.

To add detail onto my previous commit; post filtered sentences are stored into a vector of strings, so each index is for a single sentence. While the basic logic exists, I need implement a TextStatistics class which collates all relevant data points (word frequency etc.).  I may use this points to influence what words will be removed from the overall summarised text, however I feel this is outside the scope of the original hand-in I made for this application.